home page details

+++ backend/controllers/post.controller.js
@@ -1,10 +1,14 @@
 const Post = require("../models/Post");
+const mongoose = require("mongoose");
+
 const createPost = async (req, res) => {
+  const media = req.file ? req.file.filename : null;
+
   try {
     const post = new Post({
       content: req.body.content,
       user: req.user._id,
-      photo: req.body.photo,
+      photo: media,
       video: req.body.video,
       tags: req.body.tags,
       privacy: req.body.privacy,
@@ -24,7 +28,7 @@ const getAllPosts = async (req, res) => {
   try {
     console.log("Fetching posts...");
     const posts = await Post.find()
-      .populate("user", "name")
+      .populate("user", "name profilePic")
       .populate("likes", "name")
       .populate("comments.user", "name")
       .sort({ createdAt: -1 });
@@ -38,7 +42,7 @@ const getAllPosts = async (req, res) => {
 const getPostById = async (req, res) => {
   try {
     const post = await Post.findById(req.params.id)
-      .populate("user", "name")
+      .populate("user", "name profilePic")
       .populate("likes", "name")
       .populate("comments.user", "name");
 
@@ -50,6 +54,16 @@ const getPostById = async (req, res) => {
 };
 const updatePost = async (req, res) => {
   try {
+    const post = await Post.findById(req.params.id);
+
+    if (!post) return res.status(404).json({ message: "Post not found" });
+
+    if (post.user.toString() !== req.user._id.toString()) {
+      return res
+        .status(403)
+        .json({ message: "Unauthorized to edit this post" });
+    }
+
     const updatedPost = await Post.findByIdAndUpdate(
       req.params.id,
       {
@@ -62,11 +76,7 @@ const updatePost = async (req, res) => {
       { new: true }
     );
 
-    if (!updatedPost)
-      return res.status(404).json({ message: "Post not found" });
-    res
-      .status(200)
-      .json({ updatedPost: updatedPost, message: "Post updated successfully" });
+    res.status(200).json({ updatedPost, message: "Post updated successfully" });
   } catch (error) {
     res.status(500).json({ error: error.message });
   }
@@ -74,10 +84,17 @@ const updatePost = async (req, res) => {
 
 const deletePost = async (req, res) => {
   try {
-    const deletedPost = await Post.findByIdAndDelete(req.params.id);
+    const post = await Post.findById(req.params.id);
 
-    if (!deletedPost)
-      return res.status(404).json({ message: "Post not found" });
+    if (!post) return res.status(404).json({ message: "Post not found" });
+
+    if (post.user.toString() !== req.user._id.toString()) {
+      return res
+        .status(403)
+        .json({ message: "Unauthorized to delete this post" });
+    }
+
+    await Post.findByIdAndDelete(req.params.id);
     res.status(200).json({ message: "Post deleted successfully" });
   } catch (error) {
     res.status(500).json({ error: error.message });
@@ -86,11 +103,9 @@ const deletePost = async (req, res) => {
 
 const likePost = async (req, res) => {
   const id = req.params.id;
-  console.log(id);
 
   try {
     const post = await Post.findById(id);
-    console.log(id);
 
     if (!post) return res.status(404).json({ message: "Post not found" });
 
@@ -130,6 +145,424 @@ const addComment = async (req, res) => {
   }
 };
 
+const deleteComment = async (req, res) => {
+  try {
+    const post = await Post.findById(req.params.id);
+    if (!post) return res.status(404).json({ message: "Post not found" });
+    const commentId = req.params.commentId;
+    const commentIndex = post.comments.findIndex(
+      (comment) => comment._id.toString() === commentId
+    );
+
+    if (commentIndex === -1) {
+      return res.status(404).json({ message: "Comment not found" });
+    }
+    post.comments.splice(commentIndex, 1);
+    await post.save();
+
+    res.status(200).json({ message: "Comment deleted successfully" });
+  } catch (error) {
+    res.status(500).json({ error: error.message });
+  }
+};
+const likeComment = async (req, res) => {
+  const postId = req.params.id;
+  const commentId = req.params.commentId;
+  const userId = req.user._id;
+  try {
+    const post = await Post.findById(postId);
+    if (!post) return res.status(404).json({ message: "Post not found" });
+    const commentIndex = post.comments.findIndex(
+      (comment) => comment._id.toString() === commentId
+    );
+    if (commentIndex === -1)
+      return res.status(404).json({ message: "Comment not found" });
+    const comment = post.comments[commentIndex];
+    if (comment.likes.includes(userId)) {
+      comment.likes.pull(userId);
+    } else {
+      comment.likes.push(userId);
+    }
+    await post.save();
+    res.status(200).json(post);
+  } catch (error) {
+    res.status(500).json({ error: error.message });
+  }
+};
+const editComment = async (req, res) => {
+  const postId = req.params.id;
+  const commentId = req.params.commentId;
+  try {
+    const post = await Post.findById(postId);
+    if (!post) return res.status(404).json({ message: "Post not found" });
+    const commentIndex = post.comments.findIndex(
+      (comment) => comment._id.toString() === commentId
+    );
+    if (commentIndex === -1)
+      return res.status(404).json({ message: "Comment not found" });
+    const comment = post.comments[commentIndex];
+    comment.content = req.body.content;
+    await post.save();
+    res.status(200).json(post);
+  } catch (error) {
+    res.status(500).json({ error: error.message });
+  }
+};
+
+const replyComment = async (req, res) => {
+  const postId = req.params.id;
+  const commentId = req.params.commentId;
+  const userId = req.user._id;
+  try {
+    const post = await Post.findById(postId);
+    if (!post) return res.status(404).json({ message: "Post not found" });
+    const commentIndex = post.comments.findIndex(
+      (comment) => comment._id.toString() === commentId
+    );
+    if (commentIndex === -1)
+      return res.status(404).json({ message: "Comment not found" });
+    const comment = post.comments[commentIndex];
+    const reply = {
+      user: userId,
+      content: req.body.content,
+      replies: comment.replies,
+    };
+    comment.replies.push(reply);
+    await post.save();
+    res.status(200).json(post);
+  } catch (error) {
+    res.status(500).json({ error: error.message });
+  }
+};
+
+const likeReplyComment = async (req, res) => {
+  const postId = req.params.id;
+  const commentId = req.params.commentId;
+  const replyId = req.params.replyId;
+  const userId = req.user._id;
+  try {
+    const post = await Post.findById(postId);
+    if (!post) return res.status(404).json({ message: "Post not found" });
+    const commentIndex = post.comments.findIndex(
+      (comment) => comment._id.toString() === commentId
+    );
+    if (commentIndex === -1)
+      return res.status(404).json({ message: "Comment not found" });
+    const comment = post.comments[commentIndex];
+    const replyIndex = comment.replies.findIndex(
+      (reply) => reply._id.toString() === replyId
+    );
+    if (replyIndex === -1)
+      return res.status(404).json({ message: "Reply not found" });
+    const reply = comment.replies[replyIndex];
+    if (reply.likes.includes(userId)) {
+      reply.likes.pull(userId);
+    } else {
+      reply.likes.push(userId);
+    }
+    await post.save();
+    res.status(200).json(post);
+  } catch (error) {
+    res.status(500).json({ error: error.message });
+  }
+};
+const deleteReply = async (req, res) => {
+  const postId = req.params.id;
+  const commentId = req.params.commentId;
+  const replyId = req.params.replyId;
+  try {
+    const post = await Post.findById(postId);
+    if (!post) return res.status(404).json({ message: "Post not found" });
+    const commentIndex = post.comments.findIndex(
+      (comment) => comment._id.toString() === commentId
+    );
+    if (commentIndex === -1)
+      return res.status(404).json({ message: "Comment not found" });
+    const comment = post.comments[commentIndex];
+    const replyIndex = comment.replies.findIndex(
+      (reply) => reply._id.toString() === replyId
+    );
+    if (replyIndex === -1)
+      return res.status(404).json({ message: "Reply not found" });
+    comment.replies.splice(replyIndex, 1);
+    await post.save();
+    res.status(200).json(post);
+  } catch (error) {
+    res.status(500).json({ error: error.message });
+  }
+};
+
+const getCommentsByPostId = async (req, res) => {
+  const postId = req.params.id;
+  try {
+    const post = await Post.findById(postId);
+    if (!post) return res.status(404).json({ message: "Post not found" });
+    res.status(200).json(post.comments);
+  } catch (error) {
+    res.status(500).json({ error: error.message });
+  }
+};
+const getRepliesByCommentId = async (req, res) => {
+  const postId = req.params.id;
+  const commentId = req.params.commentId;
+  try {
+    const post = await Post.findById(postId);
+    if (!post) return res.status(404).json({ message: "Post not found" });
+    const commentIndex = post.comments.findIndex(
+      (comment) => comment._id.toString() === commentId
+    );
+    if (commentIndex === -1)
+      return res.status(404).json({ message: "Comment not found" });
+    const comment = post.comments[commentIndex];
+    res.status(200).json(comment.replies);
+  } catch (error) {
+    res.status(500).json({ error: error.message });
+  }
+};
+
+const editReply = async (req, res) => {
+  const postId = req.params.id;
+  const commentId = req.params.commentId;
+  const replyId = req.params.replyId;
+  try {
+    const post = await Post.findById(postId);
+    if (!post) return res.status(404).json({ message: "Post not found" });
+    const commentIndex = post.comments.findIndex(
+      (comment) => comment._id.toString() === commentId
+    );
+    if (commentIndex === -1)
+      return res.status(404).json({ message: "Comment not found" });
+    const comment = post.comments[commentIndex];
+    const replyIndex = comment.replies.findIndex(
+      (reply) => reply._id.toString() === replyId
+    );
+    if (replyIndex === -1)
+      return res.status(404).json({ message: "Reply not found" });
+    const reply = comment.replies[replyIndex];
+    reply.content = req.body.content;
+    await post.save();
+    res.status(200).json(post);
+  } catch (error) {
+    res.status(500).json({ error: error.message });
+  }
+};
+
+const replayReplayComment = async (req, res) => {
+  const postId = req.params.id;
+  const commentId = req.params.commentId;
+  const replyId = req.params.replyId;
+  const userId = req.user._id;
+  const replyContent = req.body.content;
+
+  if (!replyContent || replyContent.trim() === "") {
+    return res.status(400).json({ message: "Reply content is required" });
+  }
+
+  try {
+    const post = await Post.findById(postId);
+    if (!post) {
+      return res.status(404).json({ message: "Post not found" });
+    }
+
+    const commentIndex = post.comments.findIndex(
+      (comment) => comment._id.toString() === commentId
+    );
+    if (commentIndex === -1) {
+      return res.status(404).json({ message: "Comment not found" });
+    }
+
+    const replyIndex = post.comments[commentIndex].replies.findIndex(
+      (reply) => reply._id.toString() === replyId
+    );
+    if (replyIndex === -1) {
+      return res.status(404).json({ message: "Reply not found" });
+    }
+
+    const newReply = {
+      _id: new mongoose.Types.ObjectId(),
+      user: userId,
+      content: replyContent,
+      replies: [],
+      createdAt: new Date(),
+    };
+
+    post.comments[commentIndex].replies[replyIndex].replies.push(newReply);
+
+    await post.save();
+
+    res.status(200).json({ message: "Reply added successfully", post });
+  } catch (err) {
+    res.status(500).json({ error: err.message });
+  }
+};
+const likeReplyOnRepliedComment = async (req, res) => {
+  const postId = req.params.id;
+  const commentId = req.params.commentId;
+  const replyId = req.params.replyId;
+  const replyReplyId = req.params.replyReplyId;
+  const userId = req.user._id;
+
+  try {
+    const post = await Post.findById(postId);
+    if (!post) return res.status(404).json({ message: "Post not found" });
+
+    const commentIndex = post.comments.findIndex(
+      (comment) => comment._id.toString() === commentId
+    );
+    if (commentIndex === -1)
+      return res.status(404).json({ message: "Comment not found" });
+
+    const replyIndex = post.comments[commentIndex].replies.findIndex(
+      (reply) => reply._id.toString() === replyId
+    );
+    if (replyIndex === -1)
+      return res.status(404).json({ message: "Reply not found" });
+
+    const replyReplyIndex = post.comments[commentIndex].replies[
+      replyIndex
+    ].replies.findIndex((reply) => reply._id.toString() === replyReplyId);
+    if (replyReplyIndex === -1)
+      return res.status(404).json({ message: "Reply to reply not found" });
+
+    const replyReply =
+      post.comments[commentIndex].replies[replyIndex].replies[replyReplyIndex];
+
+    if (replyReply.likes.includes(userId)) {
+      replyReply.likes.pull(userId);
+    } else {
+      replyReply.likes.push(userId);
+    }
+
+    await post.save();
+    res.status(200).json(post);
+  } catch (error) {
+    res.status(500).json({ error: error.message });
+  }
+};
+const editReplyOnRepliedComment = async (req, res) => {
+  const postId = req.params.id;
+  const commentId = req.params.commentId;
+  const replyId = req.params.replyId;
+  const replyReplyId = req.params.replyReplyId;
+  const userId = req.user._id;
+  const newContent = req.body.content;
+
+  try {
+    const post = await Post.findById(postId);
+    if (!post) return res.status(404).json({ message: "Post not found" });
+
+    const commentIndex = post.comments.findIndex(
+      (comment) => comment._id.toString() === commentId
+    );
+    if (commentIndex === -1)
+      return res.status(404).json({ message: "Comment not found" });
+
+    const replyIndex = post.comments[commentIndex].replies.findIndex(
+      (reply) => reply._id.toString() === replyId
+    );
+    if (replyIndex === -1)
+      return res.status(404).json({ message: "Reply not found" });
+
+    const replyReplyIndex = post.comments[commentIndex].replies[
+      replyIndex
+    ].replies.findIndex((reply) => reply._id.toString() === replyReplyId);
+    if (replyReplyIndex === -1)
+      return res.status(404).json({ message: "Reply to reply not found" });
+
+    const replyReply =
+      post.comments[commentIndex].replies[replyIndex].replies[replyReplyIndex];
+
+    if (replyReply.user.toString() !== userId.toString()) {
+      return res.status(403).json({ message: "Unauthorized" });
+    }
+
+    replyReply.content = newContent;
+    await post.save();
+
+    res.status(200).json(post);
+  } catch (error) {
+    res.status(500).json({ error: error.message });
+  }
+};
+
+const deleteReplyOnRepliedComment = async (req, res) => {
+  const postId = req.params.id;
+  const commentId = req.params.commentId;
+  const replyId = req.params.replyId;
+  const replyReplyId = req.params.replyReplyId;
+  const userId = req.user._id;
+
+  try {
+    const post = await Post.findById(postId);
+    if (!post) return res.status(404).json({ message: "Post not found" });
+
+    const commentIndex = post.comments.findIndex(
+      (comment) => comment._id.toString() === commentId
+    );
+    if (commentIndex === -1)
+      return res.status(404).json({ message: "Comment not found" });
+
+    const replyIndex = post.comments[commentIndex].replies.findIndex(
+      (reply) => reply._id.toString() === replyId
+    );
+    if (replyIndex === -1)
+      return res.status(404).json({ message: "Reply not found" });
+
+    const replyReplyIndex = post.comments[commentIndex].replies[
+      replyIndex
+    ].replies.findIndex((reply) => reply._id.toString() === replyReplyId);
+    if (replyReplyIndex === -1)
+      return res.status(404).json({ message: "Reply to reply not found" });
+
+    const replyReply =
+      post.comments[commentIndex].replies[replyIndex].replies[replyReplyIndex];
+
+    if (replyReply.user.toString() !== userId.toString()) {
+      return res.status(403).json({ message: "Unauthorized" });
+    }
+
+    post.comments[commentIndex].replies[replyIndex].replies.splice(
+      replyReplyIndex,
+      1
+    );
+
+    await post.save();
+
+    res.status(200).json({ message: "Reply deleted successfully", post });
+  } catch (error) {
+    res.status(500).json({ error: error.message });
+  }
+};
+
+const sharePost = async (req, res) => {
+  try {
+    const postId = req.params.id;
+
+    const originalPost = await Post.findById(postId);
+
+    if (!originalPost) {
+      return res.status(404).json({ message: "Original post not found" });
+    }
+
+    const sharedPost = new Post({
+      content: originalPost.content,
+      photo: originalPost.photo,
+      video: originalPost.video,
+      tags: originalPost.tags,
+      user: req.user._id,
+      sharedPost: originalPost._id,
+      privacy: originalPost.privacy,
+      postType: originalPost.postType,
+    });
+
+    await sharedPost.save();
+
+    res.status(201).json({ message: "Post shared successfully", sharedPost });
+  } catch (error) {
+    res.status(500).json({ error: error.message });
+  }
+};
+
 module.exports = {
   createPost,
   getAllPosts,
@@ -138,4 +571,18 @@ module.exports = {
   deletePost,
   likePost,
   addComment,
+  deleteComment,
+  likeComment,
+  replyComment,
+  likeReplyComment,
+  deleteReply,
+  editComment,
+  getCommentsByPostId,
+  getRepliesByCommentId,
+  editReply,
+  replayReplayComment,
+  likeReplyOnRepliedComment,
+  editReplyOnRepliedComment,
+  deleteReplyOnRepliedComment,
+  sharePost,
 };
